Refactor names; Reduce number of files in the package; Add more tests
